feat: random namespace by chain name
